llamevol/__init__.py
from dotenv import load_dotenv
load_dotenv()

from .individual import Individual 
from .llamevol import LLaMEvol
from .utils import NoCodeException
from .evaluator.bo_injector import AlgorithmCritic
from .llm import LLMmanager, LLMS