pyproject.toml
[tool.poetry]
name = "llamevol"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.24.4"
ioh = "0.3.14"
scikit-learn = "1.3.0"
pandas = "2.0.3"
openai = "^1.25.2"
tqdm = "^4.66.4"
jsonlines = "^4.0.0"
smac = "^2.2.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
botorch = "^0.12.0"
gpytorch = "^1.13"
sentence-transformers = "^3.4.0"
scikit-optimize = "^0.10.2"
aisuite = {extras = ["groq", "openai"], version = "^0.1.9"}
google-genai = "^1.2.0"
cma = "3.2.2"
hebo = "^0.3.6"
cmaes = "^0.11.1"

[tool.poetry.dev-dependencies]
black = "^23.11.0"
isort = "^5.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
jupyter = "^1.0.0"
pytest-cov = "^4.1.0"
codecov = "^2.1.13"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"